requirements.txt
Babel==2.10.3
install==1.3.5
py==1.11.0
python-i18n==0.3.9
pytz==2022.2.1
PyYAML==6.0
termcolor==1.1.0
